+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG
 
 
-# 0.1.0
-- fixed brot bug (brot on empty bits string)
+# 0.1.1
+- fixed brot, rotating empty pfx.bits
+- fixed band, result has same number of bits as its first argument
+- fixed bor, result has same number of bits as its first argument
+- fixed bxor, result has same number of bits as its first argument
 
 # 0.1.0
 - initial version